Feat: "Join Notification GET API 구현"

+++ src/main/java/com/example/JustGetStartedBackEnd/API/TeamJoinNotification/Entity/JoinNotification.java
@@ -16,7 +16,7 @@ public class JoinNotification {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "notification_id")
-    private int notificationId;
+    private Long notificationId;
 
     @Column(name = "is_read")
     private boolean isRead;
